+++ app/controllers/core_induction_programmes/mentor_materials_controller.rb
@@ -1,13 +1,11 @@
 # frozen_string_literal: true
 
-class CoreInductionProgrammes::MentorMaterialsController < CoreInductionProgrammes::LessonsController
+class CoreInductionProgrammes::MentorMaterialsController < ApplicationController
   include Pundit
-
-  skip_before_action :load_course_lesson_internal
 
   after_action :verify_authorized
   before_action :authenticate_user!, except: :show
-  before_action :load_mentor_material_internal, only: %i[show edit update]
+  before_action :load_mentor_material, only: %i[show edit update]
   before_action :load_core_induction_materials, only: %i[edit update new create]
   before_action :fill_data_layer, only: %i[show edit update]
 
@@ -56,13 +54,7 @@ private
   end
 
   def load_mentor_material
-    load_course_lesson
-    id = (params[:mentor_material_id] || params[:id]).to_i - 1
-    @mentor_material = @course_lesson.mentor_materials[id]
-  end
-
-  def load_mentor_material_internal
-    load_mentor_material
+    @mentor_material = helpers.load_mentor_material_from_params
     @course_lessons = @mentor_material.core_induction_programme&.course_lessons
     authorize @mentor_material
   end
